examples/mcp-toolbox/requirements.txt
google-adk >= 1.19.0
google-genai >= 1.53.0
toolbox-core == 0.5.3
